Added general business rules

+++ Assignment3.Entities/UserRepository.cs
@@ -28,6 +28,11 @@ public class UserRepository : IUserRepository
     public Response Delete(int userId, bool force = false)
     {
         var user = _context.Users.FirstOrDefault(u => u.Id == userId);
+
+        if(user == null)
+        {
+            return Response.NotFound;
+        } 
         _context.Users.Remove(user!);
         return Response.Deleted;
     }
@@ -35,7 +40,7 @@ public class UserRepository : IUserRepository
     public UserDTO? Read(int userId)
     {
         var ReadUser = _context.Users.FirstOrDefault(u => u.Id == userId);
-        return new UserDTO(ReadUser.Id, ReadUser.Name, ReadUser.Email);
+        return ReadUser == null ? null : new UserDTO(ReadUser.Id, ReadUser.Name, ReadUser.Email);
     }
 
     public IReadOnlyCollection<UserDTO> ReadAll()
@@ -49,6 +54,7 @@ public class UserRepository : IUserRepository
     public Response Update(UserUpdateDTO user)
     {
         var entity = _context.Users.Find(user.Id);
+        if(entity == null) return Response.NotFound;
         entity.Name = user.Name;
         entity.Email = user.Email;
         _context.SaveChanges();
